pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "athena-client"
version = "1.0.0"
description = "Production-ready Python SDK for the OHDSI Athena Concepts API"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Athena-Client Core Engineering Team"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "requests>=2.25.0",
    "urllib3>=1.26.0",
    "backoff>=1.10.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "orjson>=3.9.0",
]

[project.optional-dependencies]
async = ["httpx>=0.18.0"]
pandas = ["pandas>=1.3.0"]
cli = ["click>=8.0.0", "rich>=10.0.0"]
yaml = ["pyyaml>=6.0"]
crypto = ["cryptography>=36.0.0"]
all = [
    "httpx>=0.18.0",
    "pandas>=1.3.0",
    "click>=8.0.0",
    "rich>=10.0.0",
    "pyyaml>=6.0",
    "cryptography>=36.0.0",
]
dev = [
  "ruff>=0.4.0",
  "mypy>=1.10",
  "pip-audit>=2.6",
  "pytest>=8.2",
  "pytest-cov>=5.0",
  "pytest-asyncio[mode-auto]>=0.23",
  "hypothesis>=6.103",
  "vcrpy>=6.0",
  "rich>=13.7",
  "build>=1.2",
  "twine>=5.1",
  "pytest-benchmark>=4.0",
  "bandit>=1.7.5",
  "cyclonedx-python-lib>=5.2.0",
  "types-requests",
  "types-PyYAML",
  "pandas-stubs"
]


[tool.hatch.envs.default]
features = ["dev", "all"]

[project.scripts]
athena = "athena_client.cli:main"

[project.urls]
Homepage = "https://github.com/username/athena-client"
Documentation = "https://athena-client.readthedocs.io"
Issues = "https://github.com/username/athena-client/issues"

[tool.black]
line-length = 88
target-version = ["py39"]


[tool.ruff]
target-version = "py39"
line-length = 88
lint.select = ["E", "F", "B", "I"]

[tool.mypy]
python_version = "3.9"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true

[tool.hatch.envs.default.scripts]
lint = "ruff check athena_client tests"
format = "ruff format athena_client tests"
type-check = "mypy athena_client"
test = "pytest {args}"
cov = "pytest --cov=athena_client --cov-report=term-missing {args}"
bandit-check = "bandit -c pyproject.toml -r athena_client -s B101,B404,B603"
cyclonedx = "cyclonedx-py --format json --output sbom.json"

quality = [
  "format",
  "lint",
  "type-check",
  "bandit-check",
]